Initial example template.
This is a really simple template that walks a named input
file and finds matching lines which are fed to an output file.

Nothing special, but was painful to get working. The drag and drop
wire interface isn't my favorite. And trying to figure out where
you defined things like the matching patterns didn't seem clear to me.

In fairness, I haven't read the NiFi docs much at all.